Added support for include
Complex type like ServiceCharge which has ServiceChargeDetails list will
now be serialized if for example include='ServiceDetails'. Make sure to
remove cyclical referrences which cause endless loop and therefore will
not be serialized. Do this by removing the [DataMember] attribute for
example in ServiceDetail there might be a member ServiceCharge Remove
the [DataMember] attribute so the ServiceCharge member in ServiceDetail
is not serialized which would cause endless loop.